Moved test function
