+++ govcd/edgegateway_test.go
@@ -540,23 +540,3 @@ func (vcd *TestVCD) Test_UpdateNATRule(check *C) {
 
 	check.Assert(len(edge.EdgeGateway.Configuration.EdgeGatewayServiceConfiguration.NatService.NatRule), Equals, beforeChangeNatRulesNumber)
 }
-
-func TestGetPseudoUUID(t *testing.T) {
-
-	var seen = make(map[string]int)
-
-	reUuid := regexp.MustCompile(`^[A-F0-9]{8}-[A-F0-9]{4}-[A-F0-9]{4}-[A-F0-9]{4}-[A-F0-9]{12}$`)
-	for N := 0; N < 1000; N++ {
-		uuid, _ := getPseudoUuid()
-		if !reUuid.MatchString(uuid) {
-			t.Logf("string %s doesn't look like a UUID", uuid)
-			t.Fail()
-		}
-		previous, found := seen[uuid]
-		if found {
-			t.Logf("uuid %s already in the generated list at position %d", uuid, previous)
-			t.Fail()
-		}
-		seen[uuid] = N
-	}
-}
